TensorflowCellSeperationExperiments/requirements.txt
absl-py==1.3.0
affine==2.3.1
alphashape==1.3.1
astunparse==1.6.3
attrs==22.1.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///C:/ci/chardet_1607706937985/work
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262410900/work
cryptography @ file:///C:/ci/cryptography_1616769504165/work
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.1
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
menuinst==1.4.16
networkx==2.8.7
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyopencl==2022.1.6
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
pytools==2022.1.12
PyWavelets==1.4.1
pywin32==228
rasterio==1.3.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.1
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
Shapely==1.8.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
snuggs==1.4.7
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.2.0.dev221020
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow_intel==2.10.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
trimesh==3.12.6
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.3
Werkzeug==2.2.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.14.1
wxPython==4.1.1
zipp==3.10.0
